+++ de.metas.fresh/de.metas.fresh.base/src/main/java/de/metas/fresh/picking/service/IPackingHandler.java
@@ -1,5 +1,8 @@
 package de.metas.fresh.picking.service;
 
+import de.metas.adempiere.form.IPackingItem;
+import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
+
 /*
  * #%L
  * de.metas.fresh.base
@@ -10,12 +13,12 @@ package de.metas.fresh.picking.service;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -23,12 +26,9 @@ package de.metas.fresh.picking.service;
  */
 
 
-import de.metas.adempiere.form.AbstractPackingItem;
-import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
-
 /**
  * To be used with {@link IPackingService} methods to guide their job along the road.
- * 
+ *
  * @author tsa
  *
  */
@@ -42,8 +42,8 @@ public interface IPackingHandler
 
 	/**
 	 * Called when item was packed
-	 * 
+	 *
 	 * @param itemPacked item that was packed
 	 */
-	void itemPacked(final AbstractPackingItem itemPacked);
+	void itemPacked(final IPackingItem itemPacked);
 }