+++ ReleaseNotes.md
@@ -12,7 +12,14 @@ Additional notes:
 
 The actual release notes
 
-# metasfresh 4.23.22 (upcoming)
+# metasfresh 4.23.22a (release candidate)
+
+## Features
+
+## Fixes
+ - FRESH-408 Picking Issue
+
+# metasfresh 4.23.22
 
 ## Features
  - FRESH-280 Period sorting in all Dropdowns where uses year-month numeric